requirements.txt
ipykernel
streamlit
langchain
langchain-openai
langchain-community
faiss-cpu
python-dotenv
pypdf